requirements/common.txt
Django==5.0.6
python-dotenv==1.0.1
gunicorn==22.0.0
wagtail==6.1.2
psycopg2==2.9.9
django-compressor==4.4